Implement interactive keypress
Keyboard input now highlights the corresponding
button on the UI, and README has been updated
to reflect this change.

Stylesheet has been cleaned up as part of this feature,
now hover and active brightness changes inherit the basic
button class's parameters.

All buttons have received IDs as part of this commit,
related to implementation.

The Delete key now corresponds to DEL on the calculator as well.